Añadida la transacción consultar en Usuario y TipoUsuario, con sus respectivas pruebas.

+++ servicio/src/main/java/ucab/dsw/servicio/pruebaTipoUsuario.java
@@ -4,9 +4,8 @@ import ucab.dsw.accesodatos.DaoTipoUsuario;
 import ucab.dsw.dtos.TipoUsuarioDto;
 import ucab.dsw.entidades.TipoUsuario;
 
-import javax.ws.rs.DELETE;
-import javax.ws.rs.PUT;
-import javax.ws.rs.Path;
+import javax.ws.rs.*;
+import java.util.List;
 
 public class pruebaTipoUsuario {
     @PUT
@@ -18,8 +17,8 @@ public class pruebaTipoUsuario {
         {
             DaoTipoUsuario dao = new DaoTipoUsuario();
             TipoUsuario tipoUsuario = new TipoUsuario();
-            tipoUsuario.set_descripcion("Analista");
-            tipoUsuario.set_estatus("A");
+            tipoUsuario.set_descripcion(tipoUsuarioDto.getDescripcion());
+            tipoUsuario.set_estatus(tipoUsuarioDto.getEstatus());
             TipoUsuario resul = dao.insert( tipoUsuario );
             resultado.setId( resul.get_id() );
         }
@@ -30,39 +29,59 @@ public class pruebaTipoUsuario {
         return  resultado;
     }
 
+    @GET
+    @Path("/getTipoUsuario")
+    public List<TipoUsuario> getTipoUsuario() {
+
+        List<TipoUsuario> tipoUsuario = null;
+        try {
+            DaoTipoUsuario dao = new DaoTipoUsuario();
+            tipoUsuario = dao.findAll(TipoUsuario.class);
+        } catch (Exception ex) {
+            String problema = ex.getMessage();
+        }
+        return tipoUsuario;
+    }
+
+
     @PUT
-    @Path( "/updateTipoUsuario/{id}" )
-    public TipoUsuarioDto updateTipoUsuario( TipoUsuarioDto tipoUsuarioDto)
+    @Path("/updateTipoUsuario/{id}")
+    public TipoUsuarioDto updateTipoUsuario(@PathParam("id") long id, TipoUsuarioDto tipoUsuarioDto) {
+        TipoUsuarioDto resultado = new TipoUsuarioDto();
+        try {
+            DaoTipoUsuario dao = new DaoTipoUsuario();
+            TipoUsuario tipoUsuario = dao.find(id, TipoUsuario.class);
+
+            if (tipoUsuario != null) {
+                tipoUsuario.set_descripcion(tipoUsuarioDto.getDescripcion());
+                TipoUsuario resul = dao.update(tipoUsuario);
+                resultado.setId(resul.get_id());
+            }
+
+        } catch (Exception ex) {
+            String problema = ex.getMessage();
+        }
+        return resultado;
+    }
+
+
+    @DELETE
+    @Path( "/deleteTipoUsuario/{id}" )
+    public TipoUsuarioDto deleteTipoUsuario(@PathParam("id") long id)
     {
         TipoUsuarioDto resultado = new TipoUsuarioDto();
         try
         {
             DaoTipoUsuario dao = new DaoTipoUsuario();
-            TipoUsuario tipoUsuario = new TipoUsuario(tipoUsuarioDto.getId());
-            tipoUsuario.set_descripcion(tipoUsuarioDto.getDescripcion());
-            tipoUsuario.set_estatus( tipoUsuarioDto.getEstatus() );
-            TipoUsuario resul = dao.update ( tipoUsuario );
-            resultado.setId(resul.get_id());
+            TipoUsuario tipoUsuario = dao.find(id, TipoUsuario.class);
 
+            if(tipoUsuario != null)
+            {
+                TipoUsuario resul = dao.delete(tipoUsuario);
+                resultado.setId(resul.get_id());
+            }
         }
-        catch ( Exception ex )
-        {
-            String problema = ex.getMessage();
-        }
-        return  resultado;
-    }
-
-    @DELETE
-    @Path( "/deleteTipoUsuario/{id}" )
-    public TipoUsuarioDto deleteUsuario( TipoUsuarioDto tipoUsuarioDto) {
-        TipoUsuarioDto resultado = new TipoUsuarioDto();
-        try {
-            DaoTipoUsuario dao = new DaoTipoUsuario();
-            TipoUsuario tipoUsuario = dao.find(tipoUsuarioDto.getId(), TipoUsuario.class);
-            TipoUsuario resul = dao.delete(tipoUsuario);
-            resultado.setId(resul.get_id());
-
-        } catch (Exception ex) {
+        catch (Exception ex){
             String problema = ex.getMessage();
         }
         return resultado;
